Make FeatureManager as globalActor - part 6 (Now compiles!!!)
